backend/requirements.txt
fastapi[all]
sqlalchemy 
psycopg2-binary 
python-multipart 
python-dotenv
pillow